requirements.txt
langchain
openai
elasticsearch
sentence-transformers
tiktoken
python-dotenv
